requirements.txt
django
mock
python-transifex
clint # for command line output